Update js to accomidate most functionality
Most of the main functionality is in place. Users can now play rounds
against the computer, and the number of rounds and the scores for
each player is updated on the viewport. However, the script does not
yet respond to the number of rounds reaching 5, so work continues
on this.