Add logic for ending game
Added logic to allow the player to play up to five rounds. After five
rounds, the rock, paper, and scissor buttons are disabled, and the
victor is displayed (or a draw is declared). Also, a buttons has been
added to allow the user to reset the game (by reloading the page).